Created a base class for the exposure table, replaced the previoud SpacecraftAttitudeExposureTable with a new class of the child class of the base class
